tsk-396 Move gravatar utils to contact plugin
