+++ packages/presentation/package.json
@@ -19,7 +19,6 @@
     "@hcengineering/platform-rig": "~0.6.0",
     "@typescript-eslint/eslint-plugin": "^5.30.0",
     "@typescript-eslint/parser": "^5.30.0",
-    "@types/crypto-js": "^4.1.1",
     "eslint-config-standard-with-typescript": "^21.0.1",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-node": "^11.1.0",
@@ -44,7 +43,6 @@
     "@hcengineering/image-cropper": "~0.6.0",
     "@hcengineering/client": "^0.6.3",
     "@hcengineering/setting": "~0.6.1",
-    "fast-equals": "^2.0.3",
-    "crypto-js": "^4.1.1"
+    "fast-equals": "^2.0.3"
   }
 }
