+++ packages/presentation/src/components/SelectAvatarPopup.svelte
@@ -16,12 +16,11 @@
   import { createEventDispatcher } from 'svelte'
 
   import { DropdownLabelsIntl, AnySvelteComponent, showPopup, Label } from '@hcengineering/ui'
-  import { AvatarType } from '@hcengineering/contact'
+  import { AvatarType, buildGravatarId, checkHasGravatar, getAvatarColorForId } from '@hcengineering/contact'
   import { Asset } from '@hcengineering/platform'
 
   import presentation from '..'
-  import { getAvatarTypeDropdownItems, getFileUrl, getAvatarColorForId } from '../utils'
-  import { buildGravatarId, checkHasGravatar } from '../gravatar'
+  import { getAvatarTypeDropdownItems, getFileUrl } from '../utils'
   import Card from './Card.svelte'
   import AvatarComponent from './Avatar.svelte'
   import EditAvatarPopup from './EditAvatarPopup.svelte'
