+++ packages/presentation/src/index.ts
@@ -47,7 +47,6 @@ export { connect, versionError } from './connect'
 export { default } from './plugin'
 export * from './types'
 export * from './utils'
-export * from './gravatar'
 export { presentationId }
 
 addStringsLoader(presentationId, async (lang: string) => {
